Make CDSEIS self-configurable and add SEED-capable MAKELOG to distribution.
These changes simplify installation of CDSEIS by adding a configuration script
that searches for the SAC libraries that CDSEIS needs to output SAC traces.
They also add the version of the MAKELOG program to CDSEIS that reads SEED
volumes to build log files.

See the README file for installation instructions.